Fix font rendering
Originally I used the font names as they appear in the font browser,
including the comma. The 'M' character in italic looked like it was
clipping the edges of the cell and appeared truncated. I just thought it
was a problem with the font, Kitty or my config (or a combination of the
three).

Turns out I didn't RTFM (https://sw.kovidgoyal.net/kitty/conf/#fonts)
where it says:

> To get the PostScript name for a font:

  kitty + list-fonts --psnames | grep Fira

So, locally with JetBrains Mono we get:

  kitty + list-fonts --psnames | ag JetBrains

  JetBrains Mono
      JetBrains Mono (JetBrainsMono-Regular)
      JetBrains Mono Bold (JetBrainsMono-Bold)
      JetBrains Mono Bold Italic (JetBrainsMono-BoldItalic)
      JetBrains Mono ExtraBold (JetBrainsMono-ExtraBold)
      JetBrains Mono ExtraBold Italic (JetBrainsMono-ExtraBoldItalic)
      JetBrains Mono Italic (JetBrainsMono-Italic)
      JetBrains Mono Medium (JetBrainsMono-Medium)
      JetBrains Mono Medium Italic (JetBrainsMono-MediumItalic)

No commas! This change instantly fixes all rendering issues 🎉